PDMAsyncCompletionFileNormal.cpp: Fixed harmful casts introduced in r52654.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@23212 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/PDMAsyncCompletionFileNormal.cpp
@@ -374,8 +374,8 @@ static int pdmacFileAioMgrNormalProcessTaskList(PPDMACTASKFILE pTaskHead,
                 RTFOFF OffStart = pCurr->Off & ~(RTFOFF)(512-1);
                 PDMACTASKFILETRANSFER enmTransferType = pCurr->enmTransferType;
 
-                AssertMsg((  (pCurr->enmTransferType == PDMACTASKFILETRANSFER_WRITE)
-                           || ((size_t)(OffStart + cbToTransfer) <= pEndpoint->cbFile)),
+                AssertMsg(   pCurr->enmTransferType == PDMACTASKFILETRANSFER_WRITE
+                          || (uint64_t)(OffStart + cbToTransfer) <= pEndpoint->cbFile,
                           ("Read exceeds file size OffStart=%RTfoff cbToTransfer=%d cbFile=%llu\n",
                           OffStart, cbToTransfer, pEndpoint->cbFile));
 
@@ -423,7 +423,7 @@ static int pdmacFileAioMgrNormalProcessTaskList(PPDMACTASKFILE pTaskHead,
                 if (enmTransferType == PDMACTASKFILETRANSFER_WRITE)
                 {
                     /* Grow the file if needed. */
-                    if (RT_UNLIKELY((size_t)(pCurr->Off + pCurr->DataSeg.cbSeg) > pEndpoint->cbFile))
+                    if (RT_UNLIKELY((uint64_t)(pCurr->Off + pCurr->DataSeg.cbSeg) > pEndpoint->cbFile))
                     {
                         ASMAtomicWriteU64(&pEndpoint->cbFile, pCurr->Off + pCurr->DataSeg.cbSeg);
                         RTFileSetSize(pEndpoint->File, pCurr->Off + pCurr->DataSeg.cbSeg);
@@ -737,7 +737,7 @@ int pdmacFileAioMgrNormal(RTTHREAD ThreadSelf, void *pvUser)
                         RTFOFF OffStart = pTask->Off & ~(RTFOFF)(512-1);
 
                         /* Grow the file if needed. */
-                        if (RT_UNLIKELY((size_t)(pTask->Off + pTask->DataSeg.cbSeg) > pEndpoint->cbFile))
+                        if (RT_UNLIKELY((uint64_t)(pTask->Off + pTask->DataSeg.cbSeg) > pEndpoint->cbFile))
                         {
                             ASMAtomicWriteU64(&pEndpoint->cbFile, pTask->Off + pTask->DataSeg.cbSeg);
                             RTFileSetSize(pEndpoint->File, pTask->Off + pTask->DataSeg.cbSeg);
